+++ src/core/src/op/constant.cpp
@@ -201,7 +201,7 @@ void ov::op::v0::Constant::allocate_buffer() {
 
 ov::op::v0::Constant::Constant(const element::Type& type, const ov::Shape& shape, const void* data)
     : Constant(type, shape) {
-    size_t size = ceil(shape_size(m_shape) * m_element_type.bitwidth() / 8.f);
+    size_t size = static_cast<size_t>(ceil(shape_size(m_shape) * m_element_type.bitwidth() / 8.f));
     std::memcpy(get_data_ptr_nc(), data, size);
     m_all_elements_bitwise_identical = are_all_data_elements_bitwise_identical();
 }
