+++ src/frontends/ir/src/input_model.cpp
@@ -109,7 +109,7 @@ void ParsePreProcess(pugi::xml_node& root,
 
     auto input_type = input_node->get_output_element_type(0);
     FOREACH_CHILD (chan, ppNode, "channel") {
-        int chanNo = XMLParseUtils::GetIntAttr(chan, "id", next_channel_id++);
+        int chanNo = XMLParseUtils::GetIntAttr(chan, "id", static_cast<int>(next_channel_id++));
 
         auto meanNode = chan.child("mean");
         if (!meanNode.empty()) {
